Gear Up: Integrated latest from dev.
[git-p4: depot-paths = "//depot/traktor/": change = 12327]

+++ code/Net/Replication/ReplicatorTypes.h
@@ -8,7 +8,7 @@ namespace traktor
 	namespace net
 	{
 
-typedef uint64_t handle_t;
+typedef uint8_t handle_t;
 
 	}
 }
